(fixed):revised the webpack issue after updating the node

+++ frontend/src/slices/authSlice.js
@@ -2,16 +2,24 @@ import { createSlice } from "@reduxjs/toolkit";
 
 
 // Retrieve userInfo from localStorage
-let userInfoFromStorage = null;
-try {
-  userInfoFromStorage = localStorage.getItem("userInfo");
-} catch (error) {
-  console.error("Failed to access localStorage:", error);
-}
+// let userInfoFromStorage = null;
+// try {
+//   userInfoFromStorage = localStorage.getItem("userInfo");
+// } catch (error) {
+//   console.error("Failed to access localStorage:", error);
+// }
+
+const getUserInfoFromStorage = () => {
+  if (typeof window !== "undefined") {
+    const userInfo = localStorage.getItem("userInfo");
+    return userInfo ? JSON.parse(userInfo) : null;
+  }
+  return null;
+};
 
 
 const initialState={
-  userInfo: userInfoFromStorage ? JSON.parse(userInfoFromStorage) : null,
+  userInfo: getUserInfoFromStorage(),
 }
 
 const authSlice=createSlice({
@@ -20,12 +28,16 @@ const authSlice=createSlice({
   reducers:{
     setCredentials:(state,action)=>{
       state.userInfo=action.payload;
-      localStorage.setItem("userInfo",JSON.stringify(action.payload))
+      if(typeof window !== "undefined"){
+        localStorage.setItem("userInfo",JSON.stringify(action.payload))
+      }
     },
     logout:(state,action)=>{
       state.userInfo=null;
       // localStorage.removeItem("userInfo")
-      localStorage.clear();
+      if(typeof window !== "undefined"){
+        localStorage.clear();
+      }
     }
   }
 })
